change Integer -> Double

+++ src/main/java/com/example/menu_planner/model/dtoInput/IngridientRequest.java
@@ -11,12 +11,12 @@ public record IngridientRequest(
         String name,
         @NotNull(message = "protein cannot be null")
         @PositiveOrZero(message = "protein must be greater than or equal to zero")
-        Integer protein,
+        Double protein,
         @NotNull(message = "fat cannot be null")
         @PositiveOrZero(message = "fat must be greater than or equal to zero")
-        Integer fat,
+        Double fat,
         @NotNull(message = "carbohydrates cannot be null")
         @PositiveOrZero(message = "carbohydrates must be greater than or equal to zero")
-        Integer carbohydrates
+        Double carbohydrates
 ) {
 }
